+++ traffic_ops/traffic_ops_golang/trafficvault/backends/postgres/url_sig_keys.go
@@ -66,11 +66,10 @@ func putURLSigKeys(xmlID string, tvTx *sqlx.Tx, keys tc.URLSigKeys, ctx context.
 		return err
 	}
 
-	encryptedKeyRaw, err := util.AESEncrypt(keyJSON, aesKey)
+	encryptedKey, err := util.AESEncrypt(keyJSON, aesKey)
 	if err != nil {
 		return errors.New("encrypting keys: " + err.Error())
 	}
-	encryptedKey := string(encryptedKeyRaw)
 
 	res, err := tvTx.Exec("INSERT INTO url_sig_key (deliveryservice, data) VALUES ($1, $2)", xmlID, encryptedKey)
 	if err != nil {